Backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cffi==1.15.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.1
Django==4.0
django-allauth==0.47.0
django-cors-headers==3.10.1
django-filter==21.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extensions==0.7.1
drf-nested-routers==0.93.4
gcloud==0.17.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
itypes==1.2.0
Jinja2==3.0.3
jws==0.1.3
MarkupSafe==2.0.1
oauth2client==3.0.0
oauthlib==3.1.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==1.5
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2021.3
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1